added initial attempt at docker releases for multi-arch builds

+++ build_scripts/build_sil.sh
@@ -9,15 +9,21 @@ Help()
         echo
         echo "Syntax build_sil_manager.sh [-h]"
         echo "-h        Print the help text."
+        echo "-d        When building in docker"
         echo
+
 }
 
+usingDocker=false
 
-while getopts ":h" option; do
+while getopts ":hd" option; do
         case $option in
                 h) # display help
                         Help
                         exit;;
+                d) # docker
+                        usingDocker=true
+                        ;;
                 \?) # invalid
                         echo "Error: Invalid option"
                         echo
@@ -36,19 +42,24 @@ echo "Building Ardupilot SIL and copying files to resources"
 rm -r ./src/main/resources/base/sil_manager
 mkdir ./src/main/resources/base/sil_manager
 
-session=build_sil
+if [ ${usingDocker} == true ]
+then
+  cp ./repos/ardupilot_jetson/build/sitl/bin/arducopter ./src/main/resources/base/sil_manager
+  cp ./repos/ardupilot_jetson/Tools/autotest/default_params/copter.parm ./src/main/resources/base/sil_manager
 
-tmux new-session -d -s ${session}
-tmux send-keys -t ${session} 'conda activate gust_dev' C-m
-tmux send-keys -t ${session} 'python /home/lagerprocessor/Projects/ardupilot/Tools/autotest/sim_vehicle.py --no-mavproxy' C-m
+else
+  session=build_sil
 
-sleep 10
+  tmux new-session -d -s ${session}
+  tmux send-keys -t ${session} 'conda activate gust_dev' C-m
+  tmux send-keys -t ${session} 'python /home/lagerprocessor/Projects/ardupilot/Tools/autotest/sim_vehicle.py --no-mavproxy' C-m
 
-tmux kill-session -t ${session}
+  sleep 10
 
-cp /home/lagerprocessor/Projects/ardupilot/build/sitl/bin/arducopter ./src/main/resources/base/sil_manager 
-cp /home/lagerprocessor/Projects/ardupilot/Tools/autotest/default_params/copter.parm ./src/main/resources/base/sil_manager 
-
-rm -rf ./terrain/
-rm -rf ./logs/
+  tmux kill-session -t ${session}
+  cp /home/lagerprocessor/Projects/ardupilot/build/sitl/bin/arducopter ./src/main/resources/base/sil_manager
+  cp /home/lagerprocessor/Projects/ardupilot/Tools/autotest/default_params/copter.parm ./src/main/resources/base/sil_manager
+  rm -rf ./terrain/
+  rm -rf ./logs/
+fi
